requirements.txt
# Core dependencies
faster-whisper>=0.10.0
ffmpeg-python>=0.2.0
ml-ask>=1.0.0
PyYAML>=6.0
numpy>=1.24.0
pandas>=2.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Utilities
tqdm>=4.65.0
click>=8.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for Phase 2
# speechbrain>=0.5.0  # For advanced audio analysis
# torch>=2.0.0       # Required for speechbrain